refactor: Move ProfileForm to its own directory
