+++ es-writer-extension/src/tabs/profile.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import ReactDOM from "react-dom"
 
-import ProfileForm from "../components/ProfileForm"
+import ProfileForm from "./ProfileForm"
 
 ReactDOM.render(<ProfileForm />, document.getElementById("root"))
